fixing a bunch of warnings, turning up the Release mode optimization level, and adding a vcproj.user file to make Lua script reloading not insanely slow by default when running from the debugger in Release mode (if you've opened the project before, you'll have to delete your snes9xw.vcproj.COMPUTERNAME.USERNAME.user file for this last change to take effect)

+++ trunk/lua-engine.cpp
@@ -1782,7 +1782,7 @@ static int gui_gdoverlay(lua_State *L) {
 		luaL_error(L, "bad image data or not truecolour");
 	
 	// Don't care about transparent colour
-	if (size < (11+ width*height*4))
+	if ((int)size < (11+ width*height*4))
 		luaL_error(L, "bad image data");
 	
 	const uint8* pixels = data + 11;
